🪛 Tests reflecting the user settings
