+++ esmerald/requests.py
@@ -11,6 +11,7 @@ from esmerald.typing import Void
 
 if TYPE_CHECKING:
     from esmerald.applications import Esmerald
+    from esmerald.conf.global_settings import EsmeraldAPISettings
     from esmerald.types import HTTPMethod, Receive, Scope, Send
 
 User = TypeVar("User")
@@ -47,14 +48,14 @@ class Request(StarletteRequest):
         assert (
             "global_settings" in self.scope
         ), "RequestSettingsMiddleware must be added to the middlewares"
-        return self.scope["global_settings"]
+        return cast("EsmeraldAPISettings", self.scope["global_settings"])
 
     @property
     def app_settings(self) -> Any:
         assert (
             "app_settings" in self.scope
         ), "RequestSettingsMiddleware must be added to the middlewares"
-        return self.scope["app_settings"]
+        return cast("EsmeraldAPISettings", self.scope["app_settings"])
 
     async def json(self):
         if self._json is Void:
