+++ esmerald/testclient.py
@@ -38,6 +38,7 @@ if TYPE_CHECKING:
         LifeSpanHandler,
         Middleware,
         SchedulerType,
+        SettingsType,
     )
 
 
@@ -73,41 +74,41 @@ class EsmeraldTestClient(TestClient):
 def create_client(
     routes: Union["APIGateHandler", List["APIGateHandler"]],
     *,
-    debug: bool = settings.debug,
-    app_name: Optional[str] = settings.app_name,
+    settings_config: Optional["SettingsType"] = None,
+    debug: Optional[bool] = None,
+    app_name: Optional[str] = None,
     secret_key: Optional[str] = get_random_secret_key(),
-    allowed_hosts: Optional[List[str]] = settings.allowed_hosts,
-    allow_origins: Optional[List[str]] = settings.allow_origins,
+    allowed_hosts: Optional[List[str]] = None,
+    allow_origins: Optional[List[str]] = None,
     base_url: str = "http://testserver",
     backend: "Literal['asyncio', 'trio']" = "asyncio",
     backend_options: Optional[Dict[str, Any]] = None,
-    interceptors: Optional[List["Interceptor"]] = settings.interceptors,
-    permissions: Optional[List["Permission"]] = settings.permissions,
-    dependencies: Optional["Dependencies"] = settings.dependencies,
-    middleware: Optional[List["Middleware"]] = settings.middleware,
-    csrf_config: Optional["CSRFConfig"] = settings.csrf_config,
-    exception_handlers: Optional["ExceptionHandlers"] = settings.exception_handlers,
-    openapi_config: Optional["OpenAPIConfig"] = settings.openapi_config,
-    on_shutdown: Optional[List["LifeSpanHandler"]] = settings.on_shutdown,
-    on_startup: Optional[List["LifeSpanHandler"]] = settings.on_startup,
-    cors_config: Optional["CORSConfig"] = settings.cors_config,
-    session_config: Optional["SessionConfig"] = settings.session_config,
-    scheduler_class: Optional["SchedulerType"] = settings.scheduler_class,
-    scheduler_tasks: Optional[Dict[str, str]] = settings.scheduler_tasks,
-    scheduler_configurations: Optional[
-        Dict[str, Union[str, Dict[str, str]]]
-    ] = settings.scheduler_configurations,
-    enable_scheduler: bool = settings.enable_scheduler,
+    interceptors: Optional[List["Interceptor"]] = None,
+    permissions: Optional[List["Permission"]] = None,
+    dependencies: Optional["Dependencies"] = None,
+    middleware: Optional[List["Middleware"]] = None,
+    csrf_config: Optional["CSRFConfig"] = None,
+    exception_handlers: Optional["ExceptionHandlers"] = None,
+    openapi_config: Optional["OpenAPIConfig"] = None,
+    on_shutdown: Optional[List["LifeSpanHandler"]] = None,
+    on_startup: Optional[List["LifeSpanHandler"]] = None,
+    cors_config: Optional["CORSConfig"] = None,
+    session_config: Optional["SessionConfig"] = None,
+    scheduler_class: Optional["SchedulerType"] = None,
+    scheduler_tasks: Optional[Dict[str, str]] = None,
+    scheduler_configurations: Optional[Dict[str, Union[str, Dict[str, str]]]] = None,
+    enable_scheduler: bool = None,
     raise_server_exceptions: bool = True,
-    root_path: str = "",
+    root_path: Optional[str] = "",
     static_files_config: Optional[Union["StaticFilesConfig", List["StaticFilesConfig"]]] = None,
     template_config: Optional["TemplateConfig"] = None,
-    lifespan: Optional[Callable[["Esmerald"], "AsyncContextManager"]] = settings.lifespan,
+    lifespan: Optional[Callable[["Esmerald"], "AsyncContextManager"]] = None,
     cookies: Optional[httpx._client.CookieTypes] = None,
-    redirect_slashes: bool = settings.redirect_slashes,
+    redirect_slashes: Optional[bool] = None,
 ) -> EsmeraldTestClient:
     return EsmeraldTestClient(
         app=Esmerald(
+            settings_config=settings_config,
             debug=debug,
             routes=cast("Any", routes if isinstance(routes, list) else [routes]),
             app_name=app_name,
